+++ src/main/java/core/jdbc/RowMapper.java
@@ -1,9 +1,8 @@
-package next.dao;
+package core.jdbc;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
 public interface RowMapper<T> {
-
     T mapRow(ResultSet rs) throws SQLException;
 }
