+++ src/main/java/next/dao/UserDao.java
@@ -1,5 +1,7 @@
 package next.dao;
 
+import core.jdbc.JdbcTemplate;
+import core.jdbc.RowMapperImpl;
 import next.model.User;
 
 import java.sql.SQLException;
